#173 successfully made src/econ only about treasury
